Added Changes and Routing pages
