+++ src/pages/home-tabs/Favorites.tsx
@@ -1,0 +1,38 @@
+import { 
+    IonButtons,
+      IonContent, 
+      IonHeader, 
+      IonMenuButton, 
+      IonPage, 
+      IonTitle, 
+      IonToolbar 
+  } from '@ionic/react';
+  const Favorites: React.FC = () => {
+    return (
+      <IonPage>
+        <IonHeader>
+          <IonToolbar>
+            <IonButtons slot='start'>
+              <IonMenuButton></IonMenuButton>
+            </IonButtons>
+            <IonTitle>Favorites</IonTitle>
+          </IonToolbar>
+        </IonHeader>
+        <IonContent fullscreen>
+            
+        <div
+          style={{
+            display: 'flex',
+            alignItems: 'center',
+            justifyContent: 'center',
+            height: '100%',
+          }}
+        >
+          Favorites
+        </div>
+
+        </IonContent>
+      </IonPage>
+    );
+  };
+  export default Favorites;
