+++ src/pages/menu.tsx
@@ -22,8 +22,8 @@ import Details from './Details';
 
 const Menu: React.FC = () => {
     const path = [
-        {name:'Home', url: '/it35-lab/app/home', icon: homeOutline},
-        {name:'About', url: '/it35-lab/app/about', icon: rocketOutline},
+        { name: 'Home', url: '/it35-lab/app/Home', icon: homeOutline },
+      { name: 'About', url: '/it35-lab/app/About', icon: rocketOutline },
     ]
 
     return (
@@ -55,12 +55,12 @@ const Menu: React.FC = () => {
 </IonContent>
 </IonMenu>
 <IonRouterOutlet id="main">
-<Route exact path="/it35-lab/app/home" component={Home} />
-<Route exact path="/it35-lab/app/home/details" component={Details} />
-<Route exact path="/it35-lab/app/about" component={About} />
-
-<Route exact path="/it35-lab/app">
-<Redirect to="/it35-lab/app/home"/>
+            <Route exact path="/it35-lab/app/Home" component={Home} />
+            <Route exact path="/it35-lab/app/Home/details" component={Details} />
+            <Route exact path="/it35-lab/app/About" component={About} />
+  
+            <Route exact path="/it35-lab/app">
+              <Redirect to="/it35-lab/app/Home" />
 </Route>
 </IonRouterOutlet>
